backend/requirements.txt
# FastAPI and web server
fastapi
uvicorn[standard]
python-multipart
Jinja2

# Pydantic for data validation (often a specific version is used with FastAPI)
pydantic
pydantic-settings

# Database
SQLAlchemy

# Environment variables
python-dotenv

# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# PDF and OCR processing
pandas
PyPDF2
pdfplumber
pytesseract
easyocr

# HTTP requests
requests
httpx
beautifulsoup4

# AI and ML
google-generativeai
transformers

# RAG and Vector Store
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
